rename package internal to pkg
