+++ manager/manager.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/confluentinc/confluent-kafka-go/kafka"
 	"github.com/iwanjunaid/mokabox/event"
-	"github.com/iwanjunaid/mokabox/internal/interfaces/config"
-	"github.com/iwanjunaid/mokabox/internal/interfaces/manager"
+	"github.com/iwanjunaid/mokabox/pkg/interfaces/config"
+	"github.com/iwanjunaid/mokabox/pkg/interfaces/manager"
 	"go.mongodb.org/mongo-driver/mongo"
 )
